+++ simulacion/src/main/java/cantina/ucu/Interfaces/IProducto.java
@@ -3,5 +3,5 @@ package cantina.ucu.Interfaces;
 public interface IProducto {
 
     public Double getPrecio();
-    public Double getTiempoDePreparacion();
+    public int getTiempoDePreparacion();
 }
